Replica Set 등록/삭제 API 추가 및 공통 API 변수형 수정

+++ src/main/java/kr/co/strato/adapter/k8s/statefulset/service/StatefulSetAdapterService.java
@@ -1,8 +1,14 @@
 package kr.co.strato.adapter.k8s.statefulset.service;
 
+import java.util.List;
+
+import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.stereotype.Service;
+
 import com.fasterxml.jackson.core.JsonProcessingException;
 import com.fasterxml.jackson.core.type.TypeReference;
 import com.fasterxml.jackson.databind.ObjectMapper;
+
 import feign.FeignException;
 import io.fabric8.kubernetes.api.model.apps.StatefulSet;
 import kr.co.strato.adapter.k8s.common.model.ResourceType;
@@ -12,10 +18,6 @@ import kr.co.strato.adapter.k8s.common.proxy.CommonProxy;
 import kr.co.strato.adapter.k8s.common.proxy.InNamespaceProxy;
 import kr.co.strato.global.error.exception.InternalServerException;
 import lombok.extern.slf4j.Slf4j;
-import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.stereotype.Service;
-
-import java.util.List;
 
 
 @Service
@@ -89,7 +91,7 @@ public class StatefulSetAdapterService {
      * @param statefulSetName
      * @return
      */
-    public boolean delete(Integer clusterId, String namespaceName, String statefulSetName){
+    public boolean delete(Long clusterId, String namespaceName, String statefulSetName){
         WorkloadResourceInfo reqBody = WorkloadResourceInfo.builder()
                 .kubeConfigId(clusterId)
                 .namespace(namespaceName)
